create table

+++ src/main/java/ru/skypro/homework/dto/CommentDTO.java
@@ -7,7 +7,7 @@ public class CommentDTO {
     private int author;
     private String authorImage;
     private String authorFirstName;
-    private long createdAt;
+    private int createdAt;
     private int pk;
     private String text;
